Generate an API client with Progenitor
- Validate/generate an OpenAPI schema in a test, as done in omicron
- Provide an (optional, for now) codegen'd client from the OpenAPI
