+++ lib/propolis-client/Cargo.toml
@@ -8,7 +8,9 @@ edition = "2018"
 
 [dependencies]
 propolis_types = { path = "../../crates/propolis-types" }
-reqwest = { version = "0.11", default-features = false, features = ["json", "rustls-tls"] }
+reqwest = { version = "0.11.12", default-features = false, features = ["json", "rustls-tls"] }
+base64 = "0.13"
+rand = "0.8"
 ring = "0.16"
 schemars = { version = "0.8.10", features = [ "uuid1" ] }
 serde = "1.0"
@@ -16,7 +18,14 @@ serde_json = "1.0"
 slog = { version = "2.5", features = [ "max_level_trace", "release_max_level_debug" ] }
 thiserror = "1.0"
 uuid = { version = "1.0.0", features = [ "serde", "v4" ] }
+progenitor = { git = "https://github.com/oxidecomputer/progenitor", branch = "main", optional = true }
+tokio = { version = "1.0", features = [ "net" ], optional = true }
 
 [dependencies.crucible-client-types]
 git = "https://github.com/oxidecomputer/crucible"
 rev = "144d8dafa41715e00b08a5929cc62140ff0eb561"
+
+[features]
+default = []
+generated = ["progenitor", "tokio"]
+generated-migration = ["generated"]
