+++ lib/propolis-client/src/generated.rs
@@ -1,0 +1,8 @@
+// Copyright 2022 Oxide Computer Company
+//! Experimental progenitor-generated propolis-server API client.
+
+progenitor::generate_api!(
+    spec = "../../openapi/propolis-server.json",
+    interface = Builder,
+    tags = Separate,
+);
